# mgr/dashboard: fix perf. issue when listing large amounts of buckets