@alfonsomthd 2020-09-21
> jenkins test make check