@alfonsomthd 2020-09-21
> Just as a suggestion, could you plz add some of the following details:
> [...]

@epuertat I've put more info in the description.