@alfonsomthd 2020-09-21
> > > Just as a suggestion, could you plz add some of the following details:
> > > [...]
> > 
> > 
> > @epuertat I've put more info in the description.
> 
> Thanks a lot @alfonsomthd ! Looks good to me (12k is 6 times the current brick wall). However, I'm a bit surprised by the increase in time:
> 
> * 2k -> ~3 secs
> * 10k (5x) -> 40 secs (13x)
> 
> Any clue about where this is coming? RGW REST? Dashboard API? Frontend/browser?

Mostly comes from rgw api itself. CLI equivalent for 12113 buckets:
```
[root@ceph dashboard]# time radosgw-admin bucket stats
[...]
real	0m54.991s
```