@callithea 2020-09-22
jenkins test make check